Move column transforms into tableio
